+++ tests/test_smoke.py
@@ -4,8 +4,9 @@
 
 from importlib.metadata import version as _pkg_version
 
-from specsmith import __version__
 from specsmith.config import Platform, ProjectConfig, ProjectType
+
+from specsmith import __version__
 
 
 def test_version():